メール削除の更新

+++ src/login/admin/DeleteSendMailExecuteAction.java
@@ -23,23 +23,22 @@ public class DeleteSendMailExecuteAction  extends Action {
 //		session.removeAttribute("mail_id");
 
 		//リクエストパラメータ―の取得
-		String mail_id = req.getParameter("mail_id");
-		String send_mail = req.getParameter("send_mail");
-		String ad_cd = req.getParameter("ad_cd");
+
 		Map<String, String> errors = new HashMap<>();// エラーメッセージ
 
-		SendMail sendmail = new SendMail();
-		sendmail.setSend_mail(mail_id);   //メールアドレス番号
-        sendmail.setSend_mail(send_mail); // メールアドレス
-        sendmail.setSend_mail(ad_cd);     //組織コード
-        Boolean success = smDao.deleteMail(sendmail);
+		SendMail sendmail = (SendMail)session.getAttribute("sendmail");// メールを取得
 
-        if (success) {
-            res.sendRedirect("admin_mail_delete_comp.jsp"); // 登録完了ページへ遷移
+        if (sendmail != null) {
+
+			smDao.deleteMail(sendmail);
+
+			session.removeAttribute("sendmail");
+
+            res.sendRedirect("admin_mail_delete_comp.jsp"); // 削除完了ページへ遷移
         } else {
             errors.put("database", "データベースエラーが発生しました。");
             session.setAttribute("errors", errors);
-            res.sendRedirect("admin_mail_delete.jsp"); // 入力フォームページへ戻る
+            res.sendRedirect("admin_mail_list.jsp"); // メール一覧画面へ戻る
         }
 	}
 }
